+++ httpolice/message.py
@@ -40,7 +40,8 @@ class Message(Blackboard):
 
     self_name = u'msg'
 
-    def __init__(self, version, header_entries, body, trailer_entries=None):
+    def __init__(self, version, header_entries, body, trailer_entries=None,
+                 source=None):
         super(Message, self).__init__()
         self.version = (HTTPVersion(force_unicode(version))
                         if version is not None else None)
@@ -51,6 +52,7 @@ class Message(Blackboard):
                                 for k, v in trailer_entries or []]
         self.rebuild_headers()
         self.annotations = {}
+        self.source = source
 
     @property
     def annotated_header_entries(self):