+++ httpolice/request.py
@@ -26,7 +26,7 @@ from httpolice.util.text import force_unicode
 class Request(message.Message):
 
     def __init__(self, scheme, method, target, version, header_entries,
-                 body, trailer_entries=None):
+                 body, trailer_entries=None, source=None):
         # pylint: disable=redefined-outer-name
         """
         :param scheme:
@@ -96,9 +96,15 @@ class Request(message.Message):
 
             The format is the same as for `header_entries`.
 
+        :param source:
+            If not `None`, this should be a Unicode string
+            describing where this request was obtained from.
+            For example: ``u'somefile.txt, line 2, column 4'``.
+            It may be shown in reports.
+
         """
         super(Request, self).__init__(version, header_entries, body,
-                                      trailer_entries)
+                                      trailer_entries, source)
         self.scheme = force_unicode(scheme) if scheme is not None else None
         self.method = Method(force_unicode(method))
         self.target = force_unicode(target)